requirements.txt
# Core
click>=8.1
rich>=13.0
pyyaml>=6.0
python-dotenv>=1.0

# AI SDKs
anthropic>=0.40
openai>=1.50
google-genai>=1.0

# Inbox
python-frontmatter>=1.0

# Testing
pytest>=8.0
pytest-asyncio>=0.24
